# [xamlc] fix duplicate Mono.Cecil search paths